Refactor VCS related stuff
