--- punch/vcs_use_cases/release.py
@@ -1,5 +1,0 @@
-from punch.vcs_use_cases.base import VCSUseCase
-
-
-class VCSReleaseUseCase(VCSUseCase):
-    pass
