--- punch/vcs_use_cases/tag.py
@@ -1,5 +1,0 @@
-from punch.vcs_use_cases.base import VCSUseCase
-
-
-class VCSTagUseCase(VCSUseCase):
-    pass
