+++ tests/vcs_repositories/test_vcs_repo.py
@@ -1,6 +1,8 @@
 import pytest
 import six
-from punch.vcs_repositories import base as vr, exceptions as re
+
+from punch.vcs.exceptions import RepositorySystemError
+from punch.vcs.repositories import VCSRepo
 
 if six.PY2:
     import mock
@@ -22,15 +24,15 @@ def test_init_without_program_installed(temp_empty_dir):
         else:
             mock_popen.side_effect = FileNotFoundError
 
-        with pytest.raises(re.RepositorySystemError):
-            vr.VCSRepo._set_command = _test_set_command
-            vr.VCSRepo(temp_empty_dir, mock.Mock())
+        with pytest.raises(RepositorySystemError):
+            VCSRepo._set_command = _test_set_command
+            VCSRepo(temp_empty_dir, mock.Mock())
 
 
 def test_run_without_errors(temp_empty_dir):
     with mock.patch('subprocess.check_call'):
-        vr.VCSRepo._set_command = _test_set_command
-        repo = vr.VCSRepo(temp_empty_dir, mock.Mock())
+        VCSRepo._set_command = _test_set_command
+        repo = VCSRepo(temp_empty_dir, mock.Mock())
 
     with mock.patch('subprocess.Popen') as mock_popen:
         mock_popen_obj = mock.Mock()
@@ -44,8 +46,8 @@ def test_run_without_errors(temp_empty_dir):
 
 def test_run_with_errors(temp_empty_dir):
     with mock.patch('subprocess.check_call'):
-        vr.VCSRepo._set_command = _test_set_command
-        repo = vr.VCSRepo(temp_empty_dir, mock.Mock())
+        VCSRepo._set_command = _test_set_command
+        repo = VCSRepo(temp_empty_dir, mock.Mock())
 
     with mock.patch('subprocess.Popen') as mock_popen:
         mock_popen_obj = mock.Mock()
@@ -54,7 +56,7 @@ def test_run_with_errors(temp_empty_dir):
             ("stdout".encode("utf8"), "stderr".encode("utf8"))
         mock_popen_obj.returncode = 1
 
-        with pytest.raises(vr.RepositorySystemError) as exc:
+        with pytest.raises(RepositorySystemError) as exc:
             assert repo._run([repo.command, "--help"])
 
         assert str(exc.value) == \
@@ -65,5 +67,5 @@ def test_run_with_errors(temp_empty_dir):
 def test_initialize_repo_with_global_configuration_object(
         mocker, temp_empty_dir):
     global_config = mocker.Mock()
-    repo = vr.VCSRepo(temp_empty_dir, config_obj=global_config)
+    repo = VCSRepo(temp_empty_dir, config_obj=global_config)
     assert repo.config_obj == global_config
