+++ tests/vcs_use_cases/test_tag_use_case.py
@@ -1,6 +1,6 @@
 import mock
 
-from punch.vcs_use_cases.tag import VCSTagUseCase
+from punch.vcs.use_cases import VCSTagUseCase
 
 
 def test_pre_tag():
