+++ tests/vcs_use_cases/test_use_case.py
@@ -1,6 +1,6 @@
 import mock
 
-from punch.vcs_use_cases.base import VCSUseCase
+from punch.vcs.use_cases import VCSUseCase
 
 
 def test_delegate_all():
